// src/API.ts
export type CreatePostInput = {
  id?: string | null,
  title: string,
  contents: string,
  image?: string | null,
  upvotes: number,
  downvotes: number,
};
// ...
export type Post = {
  __typename: "Post",
  id: string,
  title: string,
  contents: string,
  image?: string | null,
  upvotes: number,
  downvotes: number,
  comments?: ModelCommentConnection | null,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items:  Array<Comment | null >,
  nextToken?: string | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  postID?: string | null,
  post?: Post | null,
  content: string,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};
// ...
export type CreatePostMutation = {
  createPost?:  {
    __typename: "Post",
    id: string,
    title: string,
    contents: string,
    image?: string | null,
    upvotes: number,
    downvotes: number,
    comments?:  {
      __typename: "ModelCommentConnection",
      items:  Array< {
        __typename: "Comment",
        id: string,
        postID?: string | null,
        content: string,
        createdAt: string,
        updatedAt: string,
        owner?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    owner?: string | null,
  } | null,
};